backend/lstm_project/requirements.txt
tensorflow==2.15.0
keras-tuner==1.4.7
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
tqdm
requests
sqlalchemy
joblib
